requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.21.0
albucore==0.0.24
albumentations==2.0.8
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
attrs==24.3.0
bottle==0.13.4
bottle-websocket==0.2.9
certifi==2024.12.14
cffi==2.0.0
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
Eel==0.18.2
et_xmlfile==2.0.0
filelock==3.19.1
fonttools==4.60.0
frozenlist==1.5.0
fsspec==2025.9.0
future==1.0.0
gevent==25.9.1
gevent-websocket==0.10.1
greenlet==3.2.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.8.2
kiwisolver==1.4.9
magic-filter==1.0.12
MarkupSafe==3.0.3
matplotlib==3.10.6
mpmath==1.3.0
multidict==6.1.0
networkx==3.5
numpy==2.2.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
pillow==11.2.1
polars==1.33.1
propcache==0.2.1
psutil==7.1.0
pycparser==2.23
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
pyparsing==3.2.5
pyproject-toml==0.1.0
pysqlite3-patch==0.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32-ctypes==0.2.3
PyYAML==6.0.3
regex==2024.11.6
requests==2.32.3
router==0.1
scipy==1.16.2
simsimd==6.5.3
six==1.17.0
sniffio==1.3.1
stringzilla==4.0.15
sympy==1.14.0
tiktoken==0.8.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.203
ultralytics-thop==2.0.17
urllib3==2.3.0
yarl==1.18.3
zope.event==6.0
zope.interface==8.0.1
